+++ src/Surfaces.jl
@@ -1,0 +1,15 @@
+module Surfaces
+
+
+function expand_extrema!(a::Axis, surf::Surface)
+    ex = a[:extrema]
+    foreach(x -> expand_extrema!(ex, x), surf.surf)
+    ex
+end
+
+"For the case of representing a surface as a function of x/y... can possibly avoid allocations."
+struct SurfaceFunction <: AbstractSurface
+    f::Function
+end
+
+end